Organize Triton inference code to split workflows and models
