+++ nvtabular/inference/triton/ensemble.py
@@ -341,7 +341,7 @@ def generate_nvtabular_model(
     # copy the model file over. note that this isn't necessary with the c++ backend, but
     # does provide us to use the python backend with just changing the 'backend' parameter
     copyfile(
-        os.path.join(os.path.dirname(__file__), "model.py"),
+        os.path.join(os.path.dirname(__file__), "workflow", "model.py"),
         os.path.join(output_path, str(version), "model.py"),
     )
 
@@ -385,7 +385,7 @@ def _generate_nvtabular_config(
     and outputs to that workflow"""
     config = model_config.ModelConfig(name=name, backend=backend, max_batch_size=max_batch_size)
 
-    config.parameters["python_module"].string_value = "nvtabular.inference.triton.model"
+    config.parameters["python_module"].string_value = "nvtabular.inference.triton.workflow.model"
     config.parameters["output_model"].string_value = output_model if output_model else ""
     if sparse_max:
         # this assumes seq_length is same for each list column
@@ -533,7 +533,7 @@ def export_pytorch_model(
         cloudpickle.dump(model, o)
 
     copyfile(
-        os.path.join(os.path.dirname(__file__), "model_pt.py"),
+        os.path.join(os.path.dirname(__file__), "model", "model_pt.py"),
         os.path.join(output_path, str(version), "model.py"),
     )
